requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
sqlalchemy>=2.0.40
psycopg2-binary==2.9.10
requests==2.31.0
python-dotenv==1.0.0
pytz==2025.2

